refactor(cli): move discovery.ts → config/resolve-config.ts, remove backward-compat re-export
